fix(i3blocks): show volume for devices using the pro audio profile
For some reason, after the last Archlinux upgrade my Pixelbook speakers
now only have the Pro Audio Profile available.[1] With this profile,
the left speaker's channel is not labeled front-left--as volume-pipewire
expected--but instead it is labeled aux0.

This CL modifies the volume-pipewire block so that it can also retrieve
volume from the aux0 channel.

[1]: https://gitlab.freedesktop.org/pipewire/pipewire/-/wikis/FAQ#what-is-the-pro-audio-profile

GitOrigin-RevId: e38bfd093c1178cb0c22bfb083d1bcf233c4a5ac